+++ R/zzz.R
@@ -1,5 +1,5 @@
 .onLoad <- function(libname, pkgname){
-  reticulate::configure_environment("chromConverter")
+  env <- reticulate::configure_environment("chromConverter")
   trace_file <<- reticulate::import("aston.tracefile", delay_load = TRUE)
   pd <<- reticulate::import("pandas", delay_load = TRUE)
   csv <<- reticulate::import("csv", delay_load = TRUE)