+++ inst/shell/thermofileparser.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+mono ~/software/ThermoRawFileParser/ThermoRawFileParser.exe "$@"